requirements.txt
Flask
waitress
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
reportlab
requests
